Move connections result component one level up
