+++ client/src/components/Connections/components/InfoLoader.scss
